Move filter state into listings/index state
