+++ components/listings/index/filter.js
@@ -1,9 +1,7 @@
 import { Component } from 'react'
-import Router from 'next/router'
 
 import * as colors from '../../../constants/colors'
 import { mobileMedia } from '../../../constants/media'
-import { treatParams } from '../../../utils/filter-params.js'
 
 import FilterHeader from '../../../components/listings/index/filter/header'
 import PriceFilter from '../../../components/listings/index/filter/price'
@@ -12,257 +10,10 @@ import RoomFilter from '../../../components/listings/index/filter/rooms'
 import NeighborhoodFilter from '../../../components/listings/index/filter/neighborhoods'
 
 export default class Filter extends Component {
-  constructor(props) {
-    super(props)
-
-    const { preco_minimo, preco_maximo, area_minima, area_maxima, quartos, bairros } = props.query
-    const neighborhoods = bairros ? bairros.split('|') : []
-
-    this.state = {
-      isMobileOpen: false,
-      params: {
-        price: {
-          min: preco_minimo,
-          max: preco_maximo,
-          visible: false
-        },
-        area: {
-          min: area_minima,
-          max: area_maxima,
-          visible: false
-        },
-        rooms: {
-          value: quartos,
-          visible: false
-        },
-        neighborhoods: {
-          value: neighborhoods,
-          visible: false
-        }
-      }
-    }
-  }
-
-  handleMinPriceChange = (minPrice) => {
-    const state = this.state
-    state.params.price.min =
-      minPrice ? minPrice.value : undefined
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  handleMaxPriceChange = (maxPrice) => {
-    const state = this.state
-    state.params.price.max =
-      maxPrice ? maxPrice.value : undefined
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  handleMinAreaChange = (minArea) => {
-    const state = this.state
-    state.params.area.min =
-      minArea ? minArea.value : undefined
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  handleMaxAreaChange = (maxArea) => {
-    const state = this.state
-    state.params.area.max =
-      maxArea ? maxArea.value : undefined
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  handleRoomChange = (rooms) => {
-    const state = this.state
-    state.params.rooms.value =
-      rooms ? rooms.value : undefined
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  handleNeighborhoodChange = (value) => {
-    const state = this.state
-    state.params.neighborhoods.value = value
-    this.setState(state)
-    this.updateRoute()
-  }
-
-  updateRoute = () => {
-    const params = treatParams(this.state.params)
-
-    if (params) {
-      Router.push(`/listings/index?${params}`, `/imoveis?${params}`)
-    } else {
-      Router.push('/listings/index', '/imoveis')
-    }
-  }
-
-  resetAllParams = () => {
-    const state = this.state
-
-    state.params.price.min = undefined
-    state.params.price.max = undefined
-    state.params.area.min = undefined
-    state.params.area.max = undefined
-    state.params.rooms.value = undefined
-    state.params.neighborhoods.value = []
-
-    this.setState(state)
-
-    this.updateRoute()
-  }
-
-  toggleRoomVisibility = () => {
-    this.toggleParamVisibility('rooms')
-  }
-
-  togglePriceVisibility = () => {
-    this.toggleParamVisibility('price')
-  }
-
-  toggleAreaVisibility = () => {
-    this.toggleParamVisibility('area')
-  }
-
-  toggleNeighborhoodsVisibility = () => {
-    this.toggleParamVisibility('neighborhoods')
-  }
-
-  toggleMobilePriceVisibility = () => {
-    const { visible } = this.state.params.price
-    const state = this.state
-
-    if (visible) {
-      state.params.price.visible = false
-      state.isMobileOpen = false
-    } else {
-      state.params.price.visible = true
-      state.params.neighborhoods.visible = false
-      state.params.area.visible = false
-      state.params.rooms.visible = false
-      state.isMobileOpen = true
-    }
-
-    this.setState(state)
-  }
-
-  toggleMobileNeighborhoodsVisibility = () => {
-    const { visible } = this.state.params.neighborhoods
-    const state = this.state
-
-    if (visible) {
-      state.params.neighborhoods.visible = false
-      state.isMobileOpen = false
-    } else {
-      state.params.neighborhoods.visible = true
-      state.params.price.visible = false
-      state.params.area.visible = false
-      state.params.rooms.visible = false
-      state.isMobileOpen = true
-    }
-
-    this.setState(state)
-  }
-
-  toggleOtherMobileParams = () => {
-    const state = this.state
-    const visible = state.params.area.visible
-
-    if (visible) {
-      state.isMobileOpen = false
-      state.params.area.visible = false
-      state.params.rooms.visible = false
-    } else {
-      state.isMobileOpen = true
-      state.params.area.visible = true
-      state.params.rooms.visible = true
-    }
-
-    state.params.price.visible = false
-    state.params.neighborhoods.visible = false
-    this.setState(state)
-  }
-
-  toggleParamVisibility = (param) => {
-    const state = this.state
-    const newParamFilterVisibility = !state.params[param].visible
-
-    this.hideAllParams()
-
-    state.params[param].visible = newParamFilterVisibility
-    this.setState(state)
-  }
-
-  hideAllParams = () => {
-    const { state } = this
-    const { params } = state
-
-    Object.keys(params).map(function(key) {
-      state.params[key].visible = false
-    })
-
-    state.isMobileOpen = false
-
-    this.setState(state)
-  }
-
-  isAnyParamVisible = () => {
-    const { params } = this.state
-
-    return Object.keys(params).some(function(key) {
-      return params[key]['visible'] === true
-    })
-  }
-
-
-  getNumberOfActiveParams = () => {
-    const { price, area, rooms, neighborhoods } = this.state.params
-
-    let numberOfParams = 0
-
-    if (price.min || price.max) numberOfParams++
-    if (area.min || area.max) numberOfParams++
-    if (rooms.value) numberOfParams++
-    if (neighborhoods.value.length > 0) numberOfParams ++
-
-    return numberOfParams
-  }
-
-  renderTextForMobileMainButton = () => {
-    const numberOfParams = this.getNumberOfActiveParams()
-
-    const suffix =
-      (numberOfParams == 0) ?
-        ''
-        : ': ' + numberOfParams
-
-    return 'Filtros' + suffix
-  }
-
-  renderTextForMobileMainButton = () => {
-    const numberOfParams = this.getNumberOfActiveParams()
-
-    const suffix =
-      (numberOfParams == 0) ?
-        ''
-        : ': ' + numberOfParams
-
-    return 'Filtros' + suffix
-  }
-
-  handleOverlayClick = () => {
-    const { isMobileOpen } = this.state
-
-    if (!isMobileOpen) this.hideAllParams()
-  }
-
   render() {
     const { neighborhoodOptions } = this.props
-    const { price, area, rooms, neighborhoods } = this.state.params
-    const { isMobileOpen } = this.state
+    const { price, area, rooms, neighborhoods } = this.props.params
+    const { isMobileOpen } = this.props.params
 
     return <div className={'listings-filter-container '+ (this.isAnyParamVisible() ? 'filter-open' : '')}>
       {
